refactor: rename event tracking component
